requirements.txt
aiogram==3.8.0
psycopg2_binary==2.9.9
python_dateutil==2.9.0.post0
SQLAlchemy==2.0.31
asyncpg
